Refactor: Rename app to what beats rock
