+++ app/src/main/java/ke/don/what_beats_rock/create_itinerary/components/DraggableUtils.kt
@@ -1,4 +1,4 @@
-package ke.don.itinerar_ai.create_itinerary.components
+package ke.don.what_beats_rock.create_itinerary.components
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.gestures.detectDragGesturesAfterLongPress
@@ -18,7 +18,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.zIndex
 import kotlinx.coroutines.channels.Channel
