+++ app/src/main/java/ke/don/what_beats_rock/create_itinerary/model/ItineraryFormUiState.kt
@@ -1,4 +1,4 @@
-package ke.don.itinerar_ai.create_itinerary.model
+package ke.don.what_beats_rock.create_itinerary.model
 
 import kotlinx.serialization.Serializable
